refactor[#193] : editMyInfo -> ModifyMyInfo 로 네이밍 변경
